+++ docs/changelog.rst
@@ -5,7 +5,7 @@ Changelog
 ------------------
 
 - Nothing changed yet.
-
+- Remove unnecessary ChangeList queries to speed up export via Admin UI (#1715)
 
 3.3.4 (2023-12-09)
 ------------------